requirements.txt
altair==5.1.1
anyio==3.7.1
asyncio==3.4.3
extra-streamlit-components==0.1.56
graphviz==0.20.1
gunicorn==21.2.0
matplotlib==3.7.2 
numpy==1.23.5
pandas==2.1.1
pickle5==0.0.11
Pillow==10.0.1
pydantic==1.10.8
pydantic_core==2.6.3
pymongo==4.5.0
PyPDF2==3.0.1
requests==2.31.0
streamlit-authenticator==0.2.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-echarts==0.4.0
streamlit-embedcode==0.1.2
streamlit-extras==0.3.0
streamlit-faker==0.0.2
streamlit-image-comparison==0.0.4
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.2
streamlit-pandas-profiling==0.1.3
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
streamlit-webrtc==0.47.0
typing==3.7.4.3
